rename localStorage package and fix minor issues with Input component #146
