+++ src/features/auth/utils/clearAuthLocalStorageState.ts
@@ -3,7 +3,7 @@
  * @copyright Yury Korotovskikh <u.korotovskiy@nil.foundation>
  */
 
-import { LocalStorageAPI } from '@/packages/localStorage';
+import { LocalStorageAPI } from '@/packages/localStorageAdapter';
 
 /**
  * Clears localStorage values related to auth state.
