+++ src/packages/localStorageAdapter/LocalStorageKey.ts
