+++ src/redux/market/reducers/OrderBookReducer.ts
@@ -6,7 +6,7 @@
 import { createReducer } from '@reduxjs/toolkit';
 import type { OrderBookPriceStepType } from '@/enums';
 import type { LastOrderData, OrderBookData } from '@/models';
-import { LocalStorageAPI } from '@/packages/localStorage';
+import { LocalStorageAPI } from '@/packages/localStorageAdapter';
 import {
   UpdateOrderBookData,
   UpdateOrderBookDataError,
